+++ scripts/download_greenplum-spark-connector.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
-# Change PIVNET token
+# Change PIVNET_TOKEN, in order to use this script
+#
+#
 set -e
 set -u
 set -x
@@ -12,8 +14,8 @@ cd `dirname $0`
 # if you have pivnet cli , check cat ~/.pivnetrc
 export PIVNET_TOKEN=CHANGE_ME
 
-export LOCAL_FILE_NAME=greenplum-spark_2.11-1.0.0.jar
-export DOWNLOAD_URL=https://network.pivotal.io/api/v2/products/pivotal-gpdb/releases/7106/product_files/30352/download
+export LOCAL_FILE_NAME=greenplum-spark_2.11-1.2.0.jar
+export DOWNLOAD_URL=https://network.pivotal.io/api/v2/products/pivotal-gpdb/releases/50181/product_files/81988/download
 wget -O "$LOCAL_FILE_NAME" --post-data="" --header="Authorization: Token $PIVNET_TOKEN" $DOWNLOAD_URL
 
 cd $current