+++ CHANGELOG.md
@@ -4,4 +4,4 @@
 * Features
 
   * Examples - how to read data from Greenplum into Spark via Greenplum-Spark connector
-  
+  * How to write data from Spark into Greenplum via JDBC